Refactor island texture atlas
